% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{calculate_f_vcov}
\alias{calculate_f_vcov}
\title{helper function: wrap calculations from fit to interested covariance estimate}
\usage{
calculate_f_vcov(fit, trt_var_name, interim_data, eff_measure)
}
\description{
helper function: wrap calculations from fit to interested covariance estimate
}
