% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{calculate_V}
\alias{calculate_V}
\title{helper function: calculate V (covariance of the mean counterfactual prediction of response)}
\usage{
calculate_V(t, s, counterfact_res)
}
\description{
helper function: calculate V (covariance of the mean counterfactual prediction of response)
}
