% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_cf.R, R/treatment_effect.R
\name{confint.prediction_cf}
\alias{confint.prediction_cf}
\alias{confint.treatment_effect}
\title{Confidence interval function.}
\usage{
\method{confint}{prediction_cf}(object, parm, level = 0.95, ...)

\method{confint}{treatment_effect}(object, parm, level = 0.95, transform, ...)
}
\arguments{
\item{object}{Object to construct confidence interval.}

\item{parm}{(\code{character} or \code{integer}) Names of the parameters to construct confidence interval.}

\item{level}{(\code{numeric}) Confidence level.}

\item{...}{Not used.}

\item{transform}{(\code{function}) Transform function.}
}
\value{
A \code{matrix} of the confidence interval.
}
\description{
Confidence interval function.
}
