% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_effect.R
\name{h_diff}
\alias{h_diff}
\alias{h_jac_diff}
\alias{h_ratio}
\alias{h_jac_ratio}
\alias{h_odds_ratio}
\alias{h_jac_odds_ratio}
\title{Contrast Functions and Jacobians}
\usage{
h_diff(x)

h_jac_diff(x)

h_ratio(x)

h_jac_ratio(x)

h_odds_ratio(x)

h_jac_odds_ratio(x)
}
\arguments{
\item{x}{(\code{numeric}) Vector of values.}
}
\value{
Vector of contrasts, or matrix of jacobians.
}
\description{
Contrast Functions and Jacobians
}
\examples{
h_diff(1:3)
h_jac_ratio(1:3)
}
