% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{covadj-package}
\alias{covadj-package}
\alias{_PACKAGE}
\title{\code{covadj} Package}
\description{
\code{covadj} description
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openpharma/covadj/}
  \item Report bugs at \url{https://github.com/openpharma/covadj/issues}
}

}
\author{
\strong{Maintainer}: openpharma \email{openpharma@example.com}

}
