% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Simulated Data for Demostration}
\format{
An object of class \code{data.frame} with 500 rows and 11 columns.
}
\usage{
data
}
\description{
Simulated Data for Demostration
}
\details{
this is simulated data.
}
\keyword{datasets}
