% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy_data}
\alias{dummy_data}
\title{Dummy Trial Data with Permute-Block Randomization}
\format{
A data frame with 600 rows and 7 columns:
\describe{
\item{id}{The ID of the patients.}
\item{treatment}{The treatment assignment, "pbo", "trt1" and "trt2"}
\item{s1}{The first stratification variable, "a" and "b".}
\item{s2}{The second stratification variable, "c" and "d".}
\item{covar}{The covariate following normal distribution.}
\item{y}{The continuous response.}
\item{y_b}{The binary response.}
}
}
\source{
The data is generated by the create_dummy.R script.
}
\usage{
dummy_data
}
\description{
This dataset contains the dummy trial data with permute block randomization.
}
\keyword{dataset}
