% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_data.R
\name{find_data}
\alias{find_data}
\title{Find Data in a Fit}
\usage{
find_data(fit, ...)
}
\arguments{
\item{fit}{A fit object.}

\item{...}{Additional arguments.}
}
\value{
A data frame used in the fit.
}
\description{
Find Data in a Fit
}
