% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_anhecova.R
\name{h_adjust_pi}
\alias{h_adjust_pi}
\title{Obtain Adjustment for Proportion of Treatment Assignment}
\usage{
h_adjust_pi(pi)
}
\arguments{
\item{pi}{(\code{numeric}) vector of proportions.}
}
\value{
Numeric matrix.
}
\description{
Obtain Adjustment for Proportion of Treatment Assignment
}
\keyword{internal}
