% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_confint}
\alias{h_confint}
\title{Confidence interval calculations which are common across effect results.}
\usage{
h_confint(x, parm, level = 0.95, transform, include_se = FALSE, ...)
}
\description{
Confidence interval calculations which are common across effect results.
}
\keyword{internal}
