% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_cov_adj.R
\name{h_get_beta_estimates}
\alias{h_get_beta_estimates}
\title{Calculate Coefficient Estimates from Linear Model Input}
\usage{
h_get_beta_estimates(lm_input)
}
\arguments{
\item{lm_input}{(\code{list}) A list containing the linear model input data for each treatment arm, as returned by
\code{\link[=h_get_lm_input]{h_get_lm_input()}}.}
}
\value{
A list containing the coefficient estimates for each treatment arm.
}
\description{
This function calculates the coefficient estimates for each treatment arm from the linear model input data.
}
\keyword{internal}
