% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_anhecova.R
\name{h_get_erb}
\alias{h_get_erb}
\title{Obtain Adjustment for Covariance Matrix}
\usage{
h_get_erb(resi, group_idx, trt, pi, randomization)
}
\arguments{
\item{resi}{(\code{numeric}) vector of residuals.}

\item{group_idx}{(\code{list} of \code{integer}) index for each groups.}

\item{trt}{(\code{factor}) of treatment assignment.}

\item{pi}{(\code{numeric}) proportion of treatment assignment.}

\item{randomization}{(\code{string}) name of the randomization schema.}
}
\description{
Obtain Adjustment for Covariance Matrix
}
\keyword{internal}
