% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_cov_adj.R
\name{h_get_lm_input}
\alias{h_get_lm_input}
\title{Get Linear Model Input Data}
\usage{
h_get_lm_input(df, model)
}
\arguments{
\item{df}{(\code{data.frame}) Including the covariates needed for the \code{model}, as well as the derived outcome \code{O_hat}
and the \code{treatment} factor.}

\item{model}{(\code{formula}) The right-hand side only model formula.}
}
\value{
A list containing for each element of the \code{treatment} factor a list with the
corresponding model matrix \code{X} and the response vector \code{y}.
}
\description{
This function prepares the input data for a linear model based on the provided data frame and model formula.
}
\keyword{internal}
