% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{h_hr_coef_mat}
\alias{h_hr_coef_mat}
\title{Hazard Ratio Coefficient Matrix}
\usage{
h_hr_coef_mat(x)
}
\arguments{
\item{x}{(\code{list}) A list containing the hazard ratio estimates and their standard errors.}
}
\value{
A matrix with columns for the estimate, standard error, z-value, and p-value.
}
\description{
This function creates a coefficient matrix for the hazard ratio estimates.
}
\keyword{internal}
