% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robin_lm.R
\name{h_interaction}
\alias{h_interaction}
\title{Evaluate if Interaction Exists}
\usage{
h_interaction(formula, treatment)
}
\arguments{
\item{formula}{(\code{formula}) the formula for model fitting.}

\item{treatment}{(\code{formula}) the formula for treatment assignment.}
}
\description{
Evaluate if Interaction Exists
}
\keyword{internal}
