% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{h_log_hr_coef_mat}
\alias{h_log_hr_coef_mat}
\title{Log Hazard Ratio Coefficient Matrix}
\usage{
h_log_hr_coef_mat(x)
}
\arguments{
\item{x}{(\code{list}) A list containing the log hazard ratio estimates and their standard errors.}
}
\value{
A matrix with columns for the log hazard ratio estimate, standard error, z-value,
and p-value.
}
\description{
This function creates a coefficient matrix for the log hazard ratio estimates.
}
\keyword{internal}
