% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_effect.R
\name{h_lower_tri_idx}
\alias{h_lower_tri_idx}
\title{Lower Triangular Index}
\usage{
h_lower_tri_idx(n)
}
\arguments{
\item{n}{(\code{int}) Number of rows/columns.}
}
\value{
Matrix of lower triangular indices.
}
\description{
Lower Triangular Index
}
\keyword{internal}
