% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{h_lr_score_no_strata_no_cov}
\alias{h_lr_score_no_strata_no_cov}
\title{Log-Rank Score Function without Strata or Covariates}
\usage{
h_lr_score_no_strata_no_cov(
  theta,
  df,
  treatment,
  time,
  status,
  n = nrow(df),
  use_ties_factor = TRUE
)
}
\arguments{
\item{theta}{(\code{numeric}) The assumed log hazard ratio(s) of the second vs. the first level of the
treatment arm variable.}

\item{df}{(\code{data.frame}) The data frame containing the survival data.}

\item{treatment}{(\code{string}) The name of the treatment arm variable in \code{df}. It should be a factor with
two levels, where the first level is the reference group.}

\item{time}{(\code{string}) The name of the time variable in \code{df}, representing the survival time.}

\item{status}{(\code{string}) The name of the status variable in \code{df}, with 0 for censored and 1 for event.}

\item{n}{(\code{count}) The number of observations. Note that this can be higher than the number of rows
when used in stratified analyses computations.}

\item{use_ties_factor}{(\code{flag}) Whether to use the ties factor in the variance calculation. This is used
when calculating the score test statistic, but not when estimating the log hazard ratio.}
}
\value{
The score function value(s), with the following attributes:
\itemize{
\item \code{sigma_l2}: The variance of the log-rank statistic.
\item \code{se_theta_l}: The corresponding standard error term for the log hazard ratio.
\item \code{n}: The number of observations used in the calculation.
}
}
\description{
This function computes the standard log-rank score statistic for a survival analysis
without strata or covariates.
}
\keyword{internal}
