% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robin_mh.R
\name{h_mh_var_sato}
\alias{h_mh_var_sato}
\title{Sato Variance for the Mantel-Haenszel Risk Difference}
\usage{
h_mh_var_sato(n11k, n10k, n1k, n0k, total_weight, estimate)
}
\arguments{
\item{n11k, n10k, n1k, n0k}{(\code{matrix}) \verb{K x P} matrices of within-stratum
counts and event counts.}

\item{total_weight}{(\code{numeric}) length-\code{P} column sums of the MH weights.}

\item{estimate}{(\code{numeric}) Length-\code{P} vector of MH point estimates.}
}
\value{
Length-\code{P} numeric vector of variance estimates.
}
\description{
Sato Variance for the Mantel-Haenszel Risk Difference
}
\keyword{internal}
