% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_n_events_per_time}
\alias{h_n_events_per_time}
\title{Count Number of Events per Unique Event Time}
\usage{
h_n_events_per_time(df, time, status)
}
\arguments{
\item{df}{(\code{data.frame}) containing the survival data.}

\item{time}{(\code{string}) name of the time variable.}

\item{status}{(\code{string}) name of the status variable, where 1 indicates an event and 0 indicates censoring.}
}
\value{
A \code{data.frame} with two columns: \code{time} and \code{n_events}, where \code{n_events} is the
number of events at each time point.
}
\description{
This function counts the number of events at each unique event time point in a survival dataset.
}
\details{
If there are no events in the dataset, it returns an empty \code{data.frame}.
}
\keyword{internal}
