% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_prep_survival_vars}
\alias{h_prep_survival_vars}
\title{Prepare Survival Variables}
\usage{
h_prep_survival_vars(formula, data, treatment)
}
\arguments{
\item{formula}{(\code{formula}) with a left hand side of the form \code{Surv(time, status)}.}

\item{data}{(\code{data.frame}) containing the variables in the formula.}

\item{treatment}{(\code{string} or \code{formula}) string name of the treatment, or a formula.}
}
\value{
A list containing the following elements:
\itemize{
\item \code{time}: Name of the time variable.
\item \code{status}: Name of the status variable.
\item \code{treatment}: Name of the treatment variable.
\item \code{strata}: Name of the strata variable.
\item \code{schema}: Randomization schema.
\item \code{covariates}: Names of the covariates in the model.
\item \code{model}: A formula only including the covariates, but not treatment or strata variables.
\item \code{n_levels}: Number of treatment levels.
\item \code{levels}: Names of the treatment levels.
}
}
\description{
Prepare Survival Variables
}
\keyword{internal}
