% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{h_test_mat}
\alias{h_test_mat}
\title{Log-Rank Test Results Matrix}
\usage{
h_test_mat(x)
}
\arguments{
\item{x}{(\code{list}) A list containing the log-rank test results.}
}
\value{
A matrix with columns for the test statistic and p-value.
}
\description{
This function creates a matrix summarizing the results of the log-rank test.
}
\keyword{internal}
