% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_validate_treatment}
\alias{h_validate_treatment}
\title{Validate Treatment Variable}
\usage{
h_validate_treatment(x)
}
\arguments{
\item{x}{(\code{character})\cr or (\code{factor})\cr.}
}
\description{
Validate Treatment Variable
}
\keyword{internal}
