% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh_effect.R
\name{mh_effect_methods}
\alias{mh_effect_methods}
\alias{print.mh_effect}
\alias{table.mh_effect}
\title{S3 Methods for \code{mh_effect}}
\usage{
\method{print}{mh_effect}(x, ...)

\method{table}{mh_effect}(x, ...)
}
\arguments{
\item{x}{(\code{mh_effect}) The result of \code{\link[=robin_mh]{robin_mh()}}.}

\item{...}{Ignored additional arguments (for compatibility).}
}
\description{
S3 Methods for \code{mh_effect}
}
\section{Functions}{
\itemize{
\item \code{print(mh_effect)}: prints the \code{mh_effect} object.

\item \code{table(mh_effect)}: prints and returns invisibly the events
summary table of the \code{mh_effect} object.

}}
\examples{
x <- robin_mh(
  y_b ~ s1 + s2,
  data = subset(glm_data, treatment != "trt2"),
  treatment = treatment ~ pb(s1, s2)
)
print(x)
table(x)
}
