% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{op}
\alias{op}
\title{helper function: calculate n^(-1)f'Vf (covariance of estimated effect measure)}
\usage{
op(x, square = FALSE)
}
\description{
helper function: calculate n^(-1)f'Vf (covariance of estimated effect measure)
}
