% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_cf.R
\name{prediction_cf_methods}
\alias{prediction_cf_methods}
\alias{print.prediction_cf}
\title{S3 Methods for \code{prediction_cf}}
\usage{
\method{print}{prediction_cf}(x, level = 0.95, ...)
}
\arguments{
\item{x}{(\code{prediction_cf})\cr the obtained counter-factual prediction object.}

\item{level}{(\code{number})\cr the significance level.}
}
\value{
No return value.
}
\description{
S3 Methods for \code{prediction_cf}
}
\section{Functions}{
\itemize{
\item \code{print(prediction_cf)}: prints the prediction_cf object.

}}
\keyword{internal}
