% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{prepare_interim_vcovrc}
\alias{prepare_interim_vcovrc}
\title{helper function: prepare interim_data for calculate_f_vcov}
\usage{
prepare_interim_vcovrc(fit, trt_var_name)
}
\description{
helper function: prepare interim_data for calculate_f_vcov
}
