% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_lm.R
\name{print.summary.std_lm}
\alias{print.summary.std_lm}
\title{Print of \code{summary.std_lm} object}
\usage{
\method{print}{summary.std_lm}(x, ...)
}
\arguments{
\item{x}{(\code{summary.std_lm})\cr object.}

\item{...}{further arguments for \code{stats::printCoefmat}.}
}
\description{
Print the \code{summary.std_lm} object.
}
\keyword{internal}
