% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{randomization_schema}
\alias{randomization_schema}
\title{Randomization schema}
\format{
An object of class \code{data.frame} with 3 rows and 2 columns.
}
\usage{
randomization_schema
}
\description{
Randomization schema
}
\keyword{internal}
