% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_vcovRobinCar.R
\name{report_vcov_robincar}
\alias{report_vcov_robincar}
\title{Reporting function for vcov_robincar}
\usage{
report_vcov_robincar(x, digits = 3)
}
\arguments{
\item{x}{object returned by \code{\link{vcovRobinCar.prediction_cf}}}

\item{digits}{integer, number of decimal places of estimates in the reported table}
}
\value{
a data frame
}
\description{
Pretty reporting for the object returned by \code{\link{vcovRobinCar.prediction_cf}}.
Note, if eff_measure was not specified (i.e. NULL), no reporting table will be generated.
}
