% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robin_glm.R
\name{robin_glm}
\alias{robin_glm}
\title{Covariate adjusted glm model}
\usage{
robin_glm(
  formula,
  data,
  treatment,
  contrast = "difference",
  contrast_jac = NULL,
  vcov = vcovANHECOVA,
  family = gaussian,
  ...
)
}
\arguments{
\item{formula}{(\code{formula}) A formula of analysis.}

\item{data}{(\code{data.frame}) Input data frame.}

\item{treatment}{(\code{formula} or \code{character(1)}) A formula of treatment assignment or assignment by stratifications,
or a string name of treatment assignment.}

\item{contrast}{(\code{function} or \code{character(1)}) A function to calculate the treatment effect, or character of
"difference", "risk_ratio", "odds_ratio" for default contrasts.}

\item{contrast_jac}{(\code{function}) A function to calculate the Jacobian of the contrast function. Ignored if using
default contrasts.}

\item{vcov}{(\code{function}) A function to calculate the variance-covariance matrix of the treatment effect,
including \code{vcovHC} and \code{vcovANHECOVA}.}

\item{family}{(\code{family}) A family object of the glm model.}

\item{...}{Additional arguments passed to \code{vcov}. For finer control of glm, refer to usage of \code{treatment_effect},
\code{difference}, \code{risk_ratio}, \code{odds_ratio}.}
}
\description{
Covariate adjusted glm model
}
\examples{
robin_glm(y ~ treatment * s1, data = dummy_data, treatment = treatment ~ s1, contrast = "difference", vcov = vcovHC)
}
