% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{robin_surv}
\alias{robin_surv}
\title{Covariate Adjusted and Stratified Survival Analysis}
\usage{
robin_surv(
  formula,
  data,
  treatment,
  contrast = "hazardratio",
  test = "logrank",
  ...
)
}
\arguments{
\item{formula}{(\code{formula}) A formula of analysis, of the form
\code{Surv(time, status) ~ treatment * strata + covariates}.}

\item{data}{(\code{data.frame}) Input data frame.}

\item{treatment}{(\code{formula}) A formula of treatment assignment or assignment by stratification.}

\item{contrast}{(\code{character(1)}) The contrast statistic to be used, currently only \code{"hazardratio"}
is supported.}

\item{test}{(\code{character(1)}) The test to be used, currently only \code{"logrank"} is supported.}

\item{...}{Additional arguments passed to the survival analysis functions, in particular \code{se_method}
(please see the vignette for details).}
}
\value{
A \code{surv_effect} object containing the results of the survival analysis.
}
\description{
Calculate log-rank test as well as hazard ratio estimates for survival data, optionally adjusted
for covariates and a stratification factor.
}
\examples{
robin_surv(
  formula = Surv(time, status) ~ sex * strata + meal.cal + age,
  data = surv_data,
  treatment = sex ~ strata
)
}
\seealso{
\link{surv_effect_methods} for S3 methods.
}
