% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{robin_surv_no_strata_no_cov}
\alias{robin_surv_no_strata_no_cov}
\title{Survival Comparison without Strata or Covariates}
\usage{
robin_surv_no_strata_no_cov(vars, data, exp_level, control_level)
}
\arguments{
\item{vars}{(\code{list}) A list containing \code{levels}, \code{treatment}, and \code{covariates}.}

\item{data}{(\code{data.frame}) The data frame containing the survival data.}

\item{exp_level}{(\code{count}) Level of the experimental treatment arm.}

\item{control_level}{(\code{count}) Level of the control treatment arm.}
}
\value{
See \code{\link[=robin_surv_comparison]{robin_surv_comparison()}}.
}
\description{
This is a simple wrapper around \code{\link[=robin_surv_comparison]{robin_surv_comparison()}} called with the log-rank score function
\code{\link[=h_lr_score_no_strata_no_cov]{h_lr_score_no_strata_no_cov()}} without strata or covariates.
}
\keyword{internal}
