% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobinCar2-package.R
\docType{package}
\name{RobinCar2-package}
\alias{RobinCar2}
\alias{RobinCar2-package}
\title{\code{RobinCar2} Package}
\description{
\code{RobinCar2} implements unbiased prediction and robust inference of variance of a fit in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openpharma/RobinCar2/}
  \item Report bugs at \url{https://github.com/openpharma/RobinCar2/issues}
}

}
\author{
\strong{Maintainer}: Liming Li \email{clark.liming@gmail.com}

Authors:
\itemize{
  \item Marlena Bannick \email{mnorwood@uw.edu} (\href{https://orcid.org/0000-0001-6279-5978}{ORCID})
  \item Ting Ye
  \item Gregory Chen \email{xiangyi.gregory.chen@msd.com}
}

}
