% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_effect.R
\name{rr}
\alias{rr}
\alias{grad_rr}
\alias{or}
\alias{grad_or}
\alias{logrr}
\alias{grad_logrr}
\alias{logor}
\alias{grad_logor}
\alias{diff}
\title{Statistics of treatment effect}
\usage{
rr(x, y)

grad_rr(x, y)

or(x, y)

grad_or(x, y)

logrr(x, y)

grad_logrr(x, y)

logor(x, y)

grad_logor(x, y)

diff(x, y)
}
\description{
Statistics of treatment effect
}
