% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_glm.R
\name{std_glm}
\alias{std_glm}
\title{Standardized Generalized Linear Models}
\usage{
std_glm(formula, data, trt, ...)
}
\arguments{
\item{formula}{(\code{formula})\cr the model formula. Must include \code{trt} to indicate the treatment variable.}

\item{data}{(\code{data.frame})\cr the dataset to be used.}

\item{trt}{(\code{string})\cr the treatment variable.}

\item{...}{additional arguments for \code{glm()}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\note{
Response must be binary.
}
\examples{
std_glm(Sepal.Length ~ Species, data = iris, trt = "Species")
}
