% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_lm.R
\name{std_lm}
\alias{std_lm}
\title{Standardized Linear Models}
\usage{
std_lm(formula, data, ...)
}
\arguments{
\item{formula}{(\code{formula})\cr the model formula.}

\item{data}{(\code{data.frame})\cr the dataset to be used.}

\item{...}{additional arguments for \code{lm()}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
For linear models, we can use the conditional treatment effect to represent the marginal treatment effect
because it is collapsible. This function only conduct a \code{lm} regression and add a class \code{std_lm} to the returned object
to enable summary function.
}
\examples{
std_lm(Sepal.Length ~ Species, data = iris)
}
