% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_lm.R
\name{std_lm_methods}
\alias{std_lm_methods}
\title{Methods for \code{std_lm} Objects}
\arguments{
\item{object}{(\code{std_lm})\cr the fitted std_lm object.}
}
\value{
Depends on the method, see Details and Functions.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
\code{summary} will summarize the \code{std_lm} model with robust covariance method.
}
\examples{
object <- std_lm(Sepal.Length ~ Species, data = iris)
}
