% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_lm.R
\name{summary.std_lm}
\alias{summary.std_lm}
\title{Summary of \code{std_lm} object}
\usage{
\method{summary}{std_lm}(object, vcov = "HC3", ...)
}
\arguments{
\item{object}{(\code{std_lm})\cr object.}

\item{vcov}{(\code{string})\cr or (\code{function})\cr to obtain covariance.}

\item{...}{not used.}
}
\description{
Summarize the \code{std_lm} object.
}
\examples{
summary(object)
}
\keyword{internal}
