% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{surv_control}
\alias{surv_control}
\title{Control Survival Analysis Root Finding}
\usage{
surv_control(tol = .Machine$double.eps^0.25, maxiter = 1000, trace = 0)
}
\arguments{
\item{tol}{(\code{number}) The desired accuracy, passed to \code{\link[stats:uniroot]{stats::uniroot()}}.}

\item{maxiter}{(\code{count}) The maximum number of iterations, passed to \code{\link[stats:uniroot]{stats::uniroot()}}.}

\item{trace}{(\code{count}) Tracing level, passed to \code{\link[stats:uniroot]{stats::uniroot()}}.}
}
\value{
A named list with elements \code{tol}, \code{maxiter}, and \code{trace}.
}
\description{
Control parameters for the root-finding algorithm used by \code{\link[=robin_surv]{robin_surv()}} to estimate
the log hazard ratio.
}
