% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{surv_dat}
\alias{surv_dat}
\title{Survival Example Data}
\format{
An object of class \code{data.frame} with 228 rows and 11 columns.
}
\source{
The data is generated by the \code{create_survival.R} script.
}
\usage{
surv_dat
}
\description{
This dataset contains survival data from the \code{survival}
package's \link[survival:lung]{survival::lung} dataset, modified to include factors for
\code{sex} and \code{strata}, as well as a binary \code{status} variable which is 1
for death and 0 for censored.
}
\keyword{dataset}
