% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{survival_comparison_functions}
\alias{survival_comparison_functions}
\alias{robin_surv_no_strata_no_cov}
\alias{robin_surv_strata}
\title{Survival Comparison Functions}
\usage{
robin_surv_no_strata_no_cov(vars, data, exp_level, control_level)

robin_surv_strata(vars, data, exp_level, control_level)
}
\arguments{
\item{vars}{(\code{list}) A list containing \code{levels}, \code{treatment}, and \code{covariates}.}

\item{data}{(\code{data.frame}) The data frame containing the survival data.}

\item{exp_level}{(\code{count}) Level of the experimental treatment arm.}

\item{control_level}{(\code{count}) Level of the control treatment arm.}
}
\value{
See \code{\link[=robin_surv_comparison]{robin_surv_comparison()}}.
}
\description{
These are simple wrappers around \code{\link[=robin_surv_comparison]{robin_surv_comparison()}} called with the corresponding log-rank score
functions.
}
\section{Functions}{
\itemize{
\item \code{robin_surv_no_strata_no_cov()}: without strata and without covariates, based on
\code{\link[=h_lr_score_no_strata_no_cov]{h_lr_score_no_strata_no_cov()}}.

\item \code{robin_surv_strata()}: without strata and without covariates, based on
\code{\link[=h_lr_score_strat]{h_lr_score_strat()}}.

}}
\keyword{internal}
