% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_score.R
\name{survival_score_functions}
\alias{survival_score_functions}
\alias{h_lr_score_no_strata_no_cov}
\alias{h_lr_score_strat}
\title{Log-Rank Score Functions for Survival Analysis}
\usage{
h_lr_score_no_strata_no_cov(
  theta,
  df,
  treatment,
  time,
  status,
  n = nrow(df),
  use_ties_factor = TRUE
)

h_lr_score_strat(
  theta,
  df,
  treatment,
  time,
  status,
  strata,
  use_ties_factor = TRUE
)
}
\arguments{
\item{theta}{(\code{numeric}) The assumed log hazard ratio(s) of the second vs. the first level of the
treatment arm variable.}

\item{df}{(\code{data.frame}) The data frame containing the survival data.}

\item{treatment}{(\code{string}) The name of the treatment arm variable in \code{df}. It should be a factor with
two levels, where the first level is the reference group.}

\item{time}{(\code{string}) The name of the time variable in \code{df}, representing the survival time.}

\item{status}{(\code{string}) The name of the status variable in \code{df}, with 0 for censored and 1 for event.}

\item{n}{(\code{count}) The number of observations. Note that this can be higher than the number of rows
when used in stratified analyses computations.}

\item{use_ties_factor}{(\code{flag}) Whether to use the ties factor in the variance calculation. This is used
when calculating the score test statistic, but not when estimating the log hazard ratio.}

\item{strata}{(\code{string}) The name of the strata variable in \code{df}, which must be a factor.}

\item{covariates}{(\code{character}) The column names in \code{df} to be used for covariate adjustment.}
}
\value{
The score function value(s), with the following attributes:
\itemize{
\item \code{sigma_l2}: The variance of the log-rank statistic.
\item \code{se_theta_l}: The corresponding standard error term for the log hazard ratio.
\item \code{n}: The number of observations used in the calculation.
}
}
\description{
These functions compute the log-rank score statistics for a survival analysis.
Depending on the function, these are stratified and/or adjusted for covariates.
}
\section{Functions}{
\itemize{
\item \code{h_lr_score_no_strata_no_cov()}: without strata or covariates.

\item \code{h_lr_score_strat()}: with strata but without covariates.

}}
\keyword{internal}
