% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_anhecova.R
\name{vcovANHECOVA}
\alias{vcovANHECOVA}
\title{ANHECOVA Covariance}
\usage{
vcovANHECOVA(x, decompose = TRUE, randomization = "simple", ...)
}
\arguments{
\item{x}{(\code{prediction_cf}) Counter-factual prediction.}

\item{decompose}{(\code{flag}) whether to use decompose method to calculate the variance.}

\item{randomization}{(\code{string}) randomization method.}

\item{...}{Not used.}
}
\value{
Named covariance matrix.
}
\description{
ANHECOVA Covariance
}
