% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_anhecova.R
\name{vcovG}
\alias{vcovG}
\title{Generalized Covariance (ANHECOVA)}
\usage{
vcovG(x, decompose = TRUE, ...)
}
\arguments{
\item{x}{(\code{prediction_cf}) Counter-factual prediction.}

\item{decompose}{(\code{flag}) whether to use decompose method to calculate the variance.}

\item{...}{Not used.}
}
\value{
Named covariance matrix.
}
\description{
Generalized Covariance (ANHECOVA)
}
