% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_hc.R
\name{vcovHC}
\alias{vcovHC}
\title{Heteroskedasticity-consistent covariance matrix for predictions}
\usage{
vcovHC(x, type = "HC3", ...)
}
\description{
The heteroskedasticity-consistent covariance matrix for predictions
is obtained with \code{sandwich::vocvHC} using sandwich method.
}
