% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{vcovRobinCar.prediction_cf}
\alias{vcovRobinCar.prediction_cf}
\title{Covariance matrix of Counterfactual prediction}
\usage{
vcovRobinCar.prediction_cf(x, eff_measure = NULL, ...)
}
\arguments{
\item{eff_measure}{default is NULL, the covariance matrix of the mean counterfactual
predictions of the response. If not NULL, a valid option is "diff", "odds ratio", or "risk ratio",
in which case covariance matrix of the estimated effect measure will also be returned.}

\item{prediction_cf}{R object returned by \code{\line{predict_counterfactual}}}
}
\value{
an R object of class 'vcov_robincar', essentially a list consisting
\itemize{
\item theta: mean counterfactual prediction of the reponse, as the calculation result of
\code{\line{predict_counterfactual}}
\item V: covariance matrix of the mean counterfactual predictions of the response
\item f_vcov: covariance matrix of the estimated effect measures. It is NULL
if \code{eff_measure} is set to NULL (default)
\item n: sample size in \code{fit} of \code{\line{predict_counterfactual}}
\item eff_measure: \code{eff_measure} in the argument of this function
\item fit: \code{fit} of \code{\line{predict_counterfactual}}
\item trt_var_name: name of treatment variable in the data used in \code{fit}
of \code{\line{predict_counterfactual}}
}
}
\description{
Covariance matrix of Counterfactual prediction
}
