% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R
\name{bias}
\alias{bias}
\title{Prediction Bias}
\usage{
bias(residual, treatment, group_idx)
}
\arguments{
\item{residual}{(\code{numeric}) residuals.}

\item{treatment}{(\code{factor}) treatment.}

\item{group_idx}{(\code{character}) stratum index.}
}
\value{
Numeric matrix of bias in each stratum.
}
\description{
Obtain prediction bias within each stratum.
}
\keyword{internal}
