% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_cf.R, R/treatment_effect.R
\name{confint.prediction_cf}
\alias{confint.prediction_cf}
\alias{confint.treatment_effect}
\title{Confidence interval function.}
\usage{
\method{confint}{prediction_cf}(object, parm, level = 0.95, ...)

\method{confint}{treatment_effect}(x, parm, level = 0.95, transform, ...)
}
\arguments{
\item{transform}{\verb{(function)} A transform function.}
}
\description{
Confidence interval function.
}
