% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robin_mh.R
\name{h_mh_events_table}
\alias{h_mh_events_table}
\title{Build the Events / Patient Counts Table for \code{mh_effect}}
\usage{
h_mh_events_table(n_mat, r_mat, analysis_strata)
}
\description{
Build the Events / Patient Counts Table for \code{mh_effect}
}
\keyword{internal}
