% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robin_mh.R
\name{h_mh_var_mgr}
\alias{h_mh_var_mgr}
\title{Modified Greenland-Robins Variance for the Mantel-Haenszel Risk Difference}
\usage{
h_mh_var_mgr(n11k, n10k, n1k, n0k, weight_mat, total_weight)
}
\arguments{
\item{n11k, n10k, n1k, n0k}{(\code{matrix}) \verb{K x P} matrices of within-stratum
counts and event counts.}

\item{weight_mat}{(\code{matrix}) \verb{K x P} matrix of MH weights.}

\item{total_weight}{(\code{numeric}) length-\code{P} column sums of \code{weight_mat}.}
}
\value{
Length-\code{P} numeric vector of variance estimates.
}
\description{
Modified Greenland-Robins Variance for the Mantel-Haenszel Risk Difference
}
\keyword{internal}
