% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_couterfactual.R
\name{predict_counterfactual}
\alias{predict_counterfactual}
\title{Counterfactual Prediction}
\usage{
predict_counterfactual(fit, treatment, data, vcov, vcov_args, ...)
}
\arguments{
\item{fit}{fitted object.}

\item{treatment}{(\code{formula}) formula of form treatment ~ strata(s).}

\item{data}{(\code{data.frame}) raw dataset.}

\item{vcov}{(\code{function} or \code{character}) variance function or name.}

\item{vcov_args}{(\code{list}) additional arguments for variance function.}

\item{...}{Additional arguments for methods.}
}
\value{
Object of class \code{prediction_cf} containing of counter factual prediction.
}
\description{
Obtain counterfactual prediction of a fit.
}
