% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_couterfactual.R
\name{predict_counterfactual}
\alias{predict_counterfactual}
\title{Counterfactual Prediction}
\usage{
predict_counterfactual(fit, treatment, data)
}
\arguments{
\item{fit}{fitted object.}

\item{treatment}{(\code{formula}) formula of form treatment ~ strata(s).}

\item{data}{(\code{data.frame}) raw dataset.}
}
\value{
Numeric matrix of counter factual prediction.
}
\description{
Obtain counterfactual prediction of a fit.
}
