% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobinCar2-package.R
\docType{package}
\name{RobinCar2-package}
\alias{RobinCar2}
\alias{RobinCar2-package}
\title{\code{RobinCar2} Package}
\description{
\code{RobinCar2} implements unbiased prediction and robust inference of variance of a fit in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openpharma/RobinCar2/}
  \item Report bugs at \url{https://github.com/openpharma/RobinCar2/issues}
}

}
\author{
\strong{Maintainer}: Liming Li \email{liming.li1@astrazeneca.com} (\href{https://orcid.org/0009-0008-6870-0878}{ORCID})

Authors:
\itemize{
  \item Marlena Bannick \email{mnorwood@uw.edu} (\href{https://orcid.org/0000-0001-6279-5978}{ORCID})
  \item Dong Xi \email{Dong.Xi1@gilead.com}
  \item Ting Ye \email{tingye1@uw.edu}
  \item Yanyao Yi \email{yi_yanyao@lilly.com}
}

Other contributors:
\itemize{
  \item Gregory Chen \email{xiangyi.gregory.chen@msd.com} [contributor]
  \item Gilead Sciences, Inc. [copyright holder, funder]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item Merck Sharp & Dohme, Inc. [copyright holder, funder]
  \item AstraZeneca plc [copyright holder, funder]
  \item Eli Lilly and Company [copyright holder, funder]
  \item The University of Washington [copyright holder, funder]
}

}
