% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_effect.R
\name{treatment_effect}
\alias{treatment_effect}
\alias{difference}
\alias{risk_ratio}
\alias{odds_ratio}
\title{Treatment Effect}
\usage{
treatment_effect(
  object,
  pair = pairwise(names(object)),
  eff_measure,
  eff_jacobian = eff_jacob(eff_measure),
  ...
)

difference(object, ...)

risk_ratio(object, ...)

odds_ratio(object, ...)
}
\arguments{
\item{object}{Object from which to obtain treatment effect.}

\item{pair}{(\code{contrast}) Contrast choices.}

\item{eff_measure}{(\code{function}) Treatment effect measurement function.}

\item{eff_jacobian}{(\code{function}) Treatment effect jacobian function.}

\item{...}{Additional arguments for variance.}
}
\description{
Obtain treatment effect and variance from counter-factual prediction
}
