% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_hc.R
\name{vcovHC}
\alias{vcovHC}
\title{Heteroskedasticity-consistent covariance matrix for predictions}
\usage{
vcovHC(x, type = "HC3", ...)
}
\arguments{
\item{x}{(\code{prediction_cf}) Counter-factual prediction.}

\item{type}{(\code{character}) Type of HC covariance matrix.}

\item{...}{Additional arguments for \code{sandwich::vcovHC}.}
}
\value{
Matrix of the heteroskedasticity-consistent covariance for the predictions.
}
\description{
The heteroskedasticity-consistent covariance matrix for predictions
is obtained with \code{sandwich::vocvHC} using sandwich method.
}
